add js-modules
